requirements.txt
Flask==0.12
gunicorn==19.6.0
line-bot-sdk==1.1.0
psycopg2==2.7.3
paramiko==2.3.0
Pillow==4.0.0
